@@ -1,8 +1,8 @@
-import { Component } from '@angular/core';
+import { Component } from "@angular/core";
 
 @Component({
-  selector: 'app-root',
-  templateUrl: 'app.component.html',
-  styleUrls: ['app.component.css']
+  selector: "app-root",
+  templateUrl: "app.component.html",
+  styleUrls: ["app.component.css"]
 })
-export class AppComponent { }
+export class AppComponent {}
