@@ -1,19 +1,19 @@
-import { BrowserModule } from '@angular/platform-browser';
-import { NgModule } from '@angular/core';
-import { AppComponent } from './app.component';
+import { BrowserModule } from "@angular/platform-browser";
+import { NgModule } from "@angular/core";
+import { AppComponent } from "./app.component";
 
-import { environment } from '../environments/environment';
+import { environment } from "../environments/environment";
 
 // forms
-import { FormsModule, ReactiveFormsModule } from '@angular/forms';
+import { FormsModule, ReactiveFormsModule } from "@angular/forms";
 
 // firebase
-import { AngularFireModule } from 'angularfire2';
-import { AngularFirestoreModule } from 'angularfire2/firestore';
-import { AngularFireAuthModule } from 'angularfire2/auth';
+import { AngularFireModule } from "angularfire2";
+import { AngularFirestoreModule } from "angularfire2/firestore";
+import { AngularFireAuthModule } from "angularfire2/auth";
 
 // feature modules
-import { FeedbacksModule } from './feedbacks/feedbacks.module';
+import { FeedbacksModule } from "./feedbacks/feedbacks.module";
 
 @NgModule({
   imports: [
@@ -24,18 +24,14 @@ import { FeedbacksModule } from './feedbacks/feedbacks.module';
     ReactiveFormsModule,
 
     // firebase
-    AngularFireModule.initializeApp(environment.firebase, 'demo'), // imports firebase/app needed for everything
+    AngularFireModule.initializeApp(environment.firebase, "demo"), // imports firebase/app needed for everything
     AngularFirestoreModule, // imports firebase/firestore, only needed for database features
     AngularFireAuthModule, // imports firebase/auth, only needed for auth features
 
     // features
-    FeedbacksModule,
+    FeedbacksModule
   ],
-  declarations: [
-    AppComponent,
-  ],
-  bootstrap: [
-    AppComponent
-  ]
+  declarations: [AppComponent],
+  bootstrap: [AppComponent]
 })
-export class AppModule { }
+export class AppModule {}
