@@ -1,16 +1,15 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { async, ComponentFixture, TestBed } from "@angular/core/testing";
 
-import { FeedbackListComponent } from './feedback-list.component';
+import { FeedbackListComponent } from "./feedback-list.component";
 
-describe('FeedbackListComponent', () => {
+describe("FeedbackListComponent", () => {
   let component: FeedbackListComponent;
   let fixture: ComponentFixture<FeedbackListComponent>;
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      declarations: [ FeedbackListComponent ]
-    })
-    .compileComponents();
+      declarations: [FeedbackListComponent]
+    }).compileComponents();
   }));
 
   beforeEach(() => {
@@ -19,7 +18,7 @@ describe('FeedbackListComponent', () => {
     fixture.detectChanges();
   });
 
-  it('should create', () => {
+  it("should create", () => {
     expect(component).toBeTruthy();
   });
 });
