@@ -1,21 +1,21 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
+import { NgModule } from "@angular/core";
+import { CommonModule } from "@angular/common";
+import { BrowserAnimationsModule } from "@angular/platform-browser/animations";
 
 // forms
-import { FormsModule, ReactiveFormsModule } from '@angular/forms';
+import { FormsModule, ReactiveFormsModule } from "@angular/forms";
 
 // components
-import { FeedbackListComponent } from './feedback-list/feedback-list.component';
-import { FeedbackFormComponent } from './feedback-form/feedback-form.component';
+import { FeedbackListComponent } from "./feedback-list/feedback-list.component";
+import { FeedbackFormComponent } from "./feedback-form/feedback-form.component";
 
 // angular material
 import {
   MatListModule,
   MatInputModule,
   MatCardModule,
-  MatIconModule,
-} from '@angular/material';
+  MatIconModule
+} from "@angular/material";
 
 @NgModule({
   imports: [
@@ -29,15 +29,9 @@ import {
     MatListModule,
     MatInputModule,
     MatCardModule,
-    MatIconModule,
+    MatIconModule
   ],
-  declarations: [
-    FeedbackListComponent,
-    FeedbackFormComponent,
-  ],
-  exports: [
-    FeedbackListComponent,
-    FeedbackFormComponent,
-  ],
+  declarations: [FeedbackListComponent, FeedbackFormComponent],
+  exports: [FeedbackListComponent, FeedbackFormComponent]
 })
-export class FeedbacksModule { }
+export class FeedbacksModule {}
